fix mistakes

+++ src/main/java/student/dao/StudentDao.java
@@ -8,12 +8,19 @@ import java.util.List;
 import java.util.Optional;
 
 public interface StudentDao {
-    List<Integer> getGroupByStudentCount(int count) throws DaoException;
+    List<Integer> getGroupByStudentCount(long count) throws DaoException;
+
     List<StudentEntity> getStudentGraduateOneYear(String year) throws DaoException;
+
     List<StudentEntity> getStudentsByAlphabet() throws DaoException;
-    Optional<StudentEntity> findById(int id) throws DaoException;
-    boolean update (StudentEntity studentEntity) throws DaoException;
-    boolean create (StudentEntity studentEntity) throws DaoException;
-    boolean delete(int id) throws DaoException;
+
+    Optional<StudentEntity> findById(long id) throws DaoException;
+
+    boolean update(StudentEntity studentEntity) throws DaoException;
+
+    boolean create(StudentEntity studentEntity) throws DaoException;
+
+    boolean delete(long id) throws DaoException;
+
     void getStudentsByStatement(List<StudentEntity> result, PreparedStatement preparedStatement) throws DaoException;
 }
