+++ src/main/java/student/dao/StudentDaoImpl.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class StudentDaoImpl implements StudentDao{
+public class StudentDaoImpl implements StudentDao {
     private static final String DELETE_STUDENT_SQL = """
             DELETE FROM student
             WHERE id = ?
@@ -59,18 +59,19 @@ public class StudentDaoImpl implements StudentDao{
             FROM student
             ORDER BY second_name DESC
             """;
+
     @Override
-    public List<Integer> getGroupByStudentCount(int count) throws DaoException {
+    public List<Integer> getGroupByStudentCount(long count) throws DaoException {
         List<Integer> result = new ArrayList<>();
         try (Connection connection = ConnectionManager.get();
-             PreparedStatement prepareStatement = connection.prepareStatement(READ_GROUP)){
-            prepareStatement.setInt(1, count);
+             PreparedStatement prepareStatement = connection.prepareStatement(READ_GROUP)) {
+            prepareStatement.setLong(1, count);
             ResultSet resultSet = prepareStatement.executeQuery();
-            while (resultSet.next()){
+            while (resultSet.next()) {
                 result.add(resultSet.getInt("id"));
             }
             return result;
-        }catch (SQLException e){
+        } catch (SQLException e) {
             throw new DaoException(e);
         }
     }
@@ -79,7 +80,7 @@ public class StudentDaoImpl implements StudentDao{
     public List<StudentEntity> getStudentGraduateOneYear(String year) throws DaoException {
         List<StudentEntity> result = new ArrayList<>();
         try (Connection connection = ConnectionManager.get();
-             PreparedStatement preparedStatement = connection.prepareStatement(READ_STUDENT_GRADUATE)){
+             PreparedStatement preparedStatement = connection.prepareStatement(READ_STUDENT_GRADUATE)) {
             preparedStatement.setString(1, year);
             getStudentsByStatement(result, preparedStatement);
             return result;
@@ -87,20 +88,21 @@ public class StudentDaoImpl implements StudentDao{
             throw new DaoException(e);
         }
     }
+
     @Override
-    public Optional<StudentEntity> findById(int id) throws DaoException {
+    public Optional<StudentEntity> findById(long id) throws DaoException {
         try (Connection connection = ConnectionManager.get();
              PreparedStatement preparedStatement = connection.prepareStatement(FIND_BY_ID)) {
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
             StudentEntity student = null;
 
-            if (resultSet.next()){
+            if (resultSet.next()) {
                 student = new StudentEntity(
-                        resultSet.getInt("id"),
+                        resultSet.getLong("id"),
                         resultSet.getString("first_name"),
                         resultSet.getString("second_name"),
-                        resultSet.getInt("group_id")
+                        resultSet.getLong("group_id")
                 );
             }
             return Optional.ofNullable(student);
@@ -108,14 +110,15 @@ public class StudentDaoImpl implements StudentDao{
             throw new DaoException(e);
         }
     }
+
     @Override
-    public boolean update (StudentEntity studentEntity) throws DaoException {
+    public boolean update(StudentEntity studentEntity) throws DaoException {
         try (Connection connection = ConnectionManager.get();
              PreparedStatement preparedStatement = connection.prepareStatement(UPDATE_SQL)) {
             preparedStatement.setString(1, studentEntity.getFirstName());
             preparedStatement.setString(2, studentEntity.getSecondName());
-            preparedStatement.setInt(3, studentEntity.getGroupId());
-            preparedStatement.setInt(4, studentEntity.getId());
+            preparedStatement.setLong(3, studentEntity.getGroupId());
+            preparedStatement.setLong(4, studentEntity.getId());
 
             preparedStatement.executeUpdate();
             return true;
@@ -123,38 +126,40 @@ public class StudentDaoImpl implements StudentDao{
             throw new DaoException(e);
         }
     }
+
     @Override
-    public boolean create (StudentEntity studentEntity) throws DaoException {
-        try ( Connection connection = ConnectionManager.get();
-              PreparedStatement preparedStatement = connection.prepareStatement(CREATE_SQL, Statement.RETURN_GENERATED_KEYS)){
+    public boolean create(StudentEntity studentEntity) throws DaoException {
+        try (Connection connection = ConnectionManager.get();
+             PreparedStatement preparedStatement = connection.prepareStatement(CREATE_SQL, Statement.RETURN_GENERATED_KEYS)) {
             preparedStatement.setString(1, studentEntity.getFirstName());
             preparedStatement.setString(2, studentEntity.getSecondName());
-            preparedStatement.setInt(3, studentEntity.getGroupId());
+            preparedStatement.setLong(3, studentEntity.getGroupId());
 
             preparedStatement.executeUpdate();
 
             ResultSet generatedKeys = preparedStatement.getGeneratedKeys();
-            if (generatedKeys.next()){
-                studentEntity.setId(generatedKeys.getInt("id"));
+            if (generatedKeys.next()) {
+                studentEntity.setId(generatedKeys.getLong("id"));
             }
             return true;
-        }catch (SQLException e){
+        } catch (SQLException e) {
             throw new DaoException(e);
         }
     }
+
     @Override
-    public boolean delete(int id) throws DaoException {
+    public boolean delete(long id) throws DaoException {
         try (Connection connection = ConnectionManager.get();
              PreparedStatement deleteRecordBookStatement = connection.prepareStatement(DELETE_RECORD_BOOK_SQL);
-             PreparedStatement deleteStudentStatement = connection.prepareStatement(DELETE_STUDENT_SQL)){
+             PreparedStatement deleteStudentStatement = connection.prepareStatement(DELETE_STUDENT_SQL)) {
 
-            deleteRecordBookStatement.setInt(1, id);
-            deleteStudentStatement.setInt(1, id);
+            deleteRecordBookStatement.setLong(1, id);
+            deleteStudentStatement.setLong(1, id);
 
             deleteRecordBookStatement.executeUpdate();
             deleteStudentStatement.executeUpdate();
             return true;
-        }catch (SQLException sqlException){
+        } catch (SQLException sqlException) {
             throw new DaoException(sqlException);
         }
     }
@@ -163,9 +168,9 @@ public class StudentDaoImpl implements StudentDao{
     public List<StudentEntity> getStudentsByAlphabet() throws DaoException {
         List<StudentEntity> result = new ArrayList<>();
         try (Connection connection = ConnectionManager.get();
-             PreparedStatement preparedStatement = connection.prepareStatement(READ_STUDENT_BY_ALPHABET)){
+             PreparedStatement preparedStatement = connection.prepareStatement(READ_STUDENT_BY_ALPHABET)) {
             getStudentsByStatement(result, preparedStatement);
-        }catch (SQLException sqlException){
+        } catch (SQLException sqlException) {
             throw new DaoException(sqlException);
         }
         return result;
@@ -174,16 +179,16 @@ public class StudentDaoImpl implements StudentDao{
     public void getStudentsByStatement(List<StudentEntity> result, PreparedStatement preparedStatement) throws DaoException {
         try {
             ResultSet resultSet = preparedStatement.executeQuery();
-            while (resultSet.next()){
+            while (resultSet.next()) {
                 StudentEntity student = new StudentEntity(
-                        resultSet.getInt("id"),
+                        resultSet.getLong("id"),
                         resultSet.getString("first_name"),
                         resultSet.getString("second_name"),
-                        resultSet.getInt("group_id")
+                        resultSet.getLong("group_id")
                 );
                 result.add(student);
             }
-        }catch (SQLException sqlException) {
+        } catch (SQLException sqlException) {
             throw new DaoException(sqlException);
         }
     }
